// frontend/src/api/generated/models/inventory-location-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UnitType } from './unit-type';

/**
 * 
 * @export
 * @interface InventoryLocationItem
 */
export interface InventoryLocationItem {
    /**
     * 在庫レコードID
     * @type {number}
     * @memberof InventoryLocationItem
     */
    'id': number;
    /**
     * ロケーションID
     * @type {number}
     * @memberof InventoryLocationItem
     */
    'locationId': number;
    /**
     * ロケーションコード
     * @type {string}
     * @memberof InventoryLocationItem
     */
    'locationCode': string;
    /**
     * 商品ID
     * @type {number}
     * @memberof InventoryLocationItem
     */
    'productId': number;
    /**
     * 商品コード
     * @type {string}
     * @memberof InventoryLocationItem
     */
    'productCode': string;
    /**
     * 商品名
     * @type {string}
     * @memberof InventoryLocationItem
     */
    'productName': string;
    /**
     * 
     * @type {UnitType}
     * @memberof InventoryLocationItem
     */
    'unitType': UnitType;
    /**
     * ロット番号
     * @type {string}
     * @memberof InventoryLocationItem
     */
    'lotNumber'?: string | null;
    /**
     * 賞味/使用期限
     * @type {string}
     * @memberof InventoryLocationItem
     */
    'expiryDate'?: string | null;
    /**
     * 在庫数量
     * @type {number}
     * @memberof InventoryLocationItem
     */
    'quantity': number;
    /**
     * 引当数量
     * @type {number}
     * @memberof InventoryLocationItem
     */
    'allocatedQty': number;
    /**
     * 有効在庫数（quantity - allocatedQty）
     * @type {number}
     * @memberof InventoryLocationItem
     */
    'availableQty': number;
    /**
     * 最終更新日時
     * @type {string}
     * @memberof InventoryLocationItem
     */
    'updatedAt': string;
}